pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cicd_tools"
version = "0.1.0"
description = "A flexible framework for development tasks including project creation, testing, building, and deployment"
readme = "README.md"
authors = [
    {name = "CICD Tools Team"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
]
requires-python = ">=3.8"
dependencies = [
    "questionary==2.1.0",
    "copier==9.6.0",
    "pyyaml==6.0.2",
    "click==8.1.8",
    "python-env-manager==0.1.0",
    "rich==13.9.4"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "bump2version>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.scripts]
cicd_tools = "cicd_tools.cli:main"

[tool.setuptools]
packages = ["cicd_tools"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
# Test coverage disabled
# addopts = "--cov=cicd_tools --cov-report=term-missing"
